+++ opentelemetry-jaeger/src/lib.rs
@@ -540,7 +540,11 @@ mod propagator {
 
     impl TextMapPropagator for Propagator {
         fn inject_context(&self, cx: &Context, injector: &mut dyn Injector) {
-            let span = cx.span();
+            let span = match cx.span() {
+                Some(span) => span,
+                None => return,
+            };
+
             let span_context = span.span_context();
             if span_context.is_valid() {
                 let flag: u8 = if span_context.is_sampled() {
@@ -705,7 +709,7 @@ mod propagator {
                 let mut map: HashMap<String, String> = HashMap::new();
                 map.set(context_key, format!("{}:{}:0:{}", trace_id, span_id, flag));
                 let context = propagator.extract(&map);
-                assert_eq!(context.span().span_context(), &expected);
+                assert_eq!(context.span().unwrap().span_context(), &expected);
             }
         }
 
@@ -728,7 +732,10 @@ mod propagator {
             let map: HashMap<String, String> = HashMap::new();
             let propagator = Propagator::new();
             let context = propagator.extract(&map);
-            assert_eq!(context.span().span_context(), &SpanContext::empty_context())
+            assert_eq!(
+                context.span().unwrap().span_context(),
+                &SpanContext::empty_context()
+            )
         }
 
         #[test]
@@ -749,7 +756,7 @@ mod propagator {
                     format!("{}:{}:0:{}", trace_id, span_id, flag),
                 );
                 let context = propagator.extract(&map);
-                assert_eq!(context.span().span_context(), &expected);
+                assert_eq!(context.span().unwrap().span_context(), &expected);
             }
         }
 
@@ -762,7 +769,10 @@ mod propagator {
             );
             let propagator = Propagator::new();
             let context = propagator.extract(&map);
-            assert_eq!(context.span().span_context(), &SpanContext::empty_context());
+            assert_eq!(
+                context.span().unwrap().span_context(),
+                &SpanContext::empty_context()
+            );
         }
 
         #[test]
@@ -774,7 +784,10 @@ mod propagator {
             );
             let propagator = Propagator::new();
             let context = propagator.extract(&map);
-            assert_eq!(context.span().span_context(), &SpanContext::empty_context());
+            assert_eq!(
+                context.span().unwrap().span_context(),
+                &SpanContext::empty_context()
+            );
         }
 
         #[test]
@@ -787,7 +800,7 @@ mod propagator {
             let propagator = Propagator::new();
             let context = propagator.extract(&map);
             assert_eq!(
-                context.span().span_context(),
+                context.span().unwrap().span_context(),
                 &SpanContext::new(
                     TraceId::from_u128(TRACE_ID),
                     SpanId::from_u64(SPAN_ID),